Migrated the logic for getting latest SAST scan results (migrated from ResultsService to SastScanner). Refactoring.
