+++ src/main/java/com/checkmarx/flow/service/SCAScanner.java
@@ -4,9 +4,7 @@ import com.checkmarx.flow.config.FlowProperties;
 import com.checkmarx.sdk.config.ScaProperties;
 import com.checkmarx.sdk.dto.ScanResults;
 import com.checkmarx.sdk.dto.ast.ASTResultsWrapper;
-import com.checkmarx.sdk.service.AstClient;
 import com.cx.restclient.ScaClientImpl;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -15,12 +13,10 @@ import java.util.Optional;
 
 @Service
 @Slf4j
-public class SCAScanner extends  AbstractASTScanner{
-
+public class SCAScanner extends AbstractASTScanner {
     public SCAScanner(ScaClientImpl astClient, FlowProperties flowProperties) {
         super(astClient, flowProperties, ScaProperties.CONFIG_PREFIX);
     }
-
 
     @Override
     protected ScanResults toScanResults(ASTResultsWrapper internalResults) {
@@ -29,13 +25,8 @@ public class SCAScanner extends  AbstractASTScanner{
                 .build();
     }
 
-
     @Override
     protected String getScanId(ASTResultsWrapper internalResults) {
-
         return Optional.ofNullable(internalResults.getScaResults().getScanId()).orElse("");
     }
-
-
-
 }
