+++ src/main/java/com/checkmarx/flow/service/VulnerabilityScanner.java
@@ -3,18 +3,17 @@ package com.checkmarx.flow.service;
 import com.checkmarx.flow.dto.ScanRequest;
 import com.checkmarx.sdk.dto.ScanResults;
 
+import javax.annotation.CheckForNull;
 import java.io.File;
 
-import javax.annotation.CheckForNull;
-
 public interface VulnerabilityScanner {
-
-    
     @CheckForNull
     ScanResults scan(ScanRequest scanRequest);
 
     @CheckForNull
-	ScanResults scanCli(ScanRequest request, String scanType, File... files);
+    ScanResults scanCli(ScanRequest request, String scanType, File... files);
+
+    ScanResults getLatestScanResults(ScanRequest request);
 
     boolean isEnabled();
 }
