+++ src/test/java/com/checkmarx/flow/cucumber/integration/azure/publishing/github2ado/Github2AdoSteps.java
@@ -307,8 +307,7 @@ public class Github2AdoSteps {
                 null,
                 null,
                 emailService,
-                cxProperties,
-                flowProperties));
+                cxProperties));
     }
 
     private ScanResults createFakeResults() {
